refactor: move model_via_field_ty into codegen_support::via
It is only called from macro-generated code, so it belongs alongside the
other codegen helpers rather than in the public `schema` module.
